move prisma output to services
