+++ package.json
@@ -8,9 +8,11 @@
     "generate:organizations": "cd services/organizations/ && npx prisma generate",
     "generate:activities": "cd services/activities/ && npx prisma generate",
     "generate:people": "cd services/people/ && npx prisma generate",
+    "generate": "concurrently npm:generate:*",
     "push:organizations": "cd services/organizations/ && npx prisma db push",
     "push:activities": "cd services/activities/ && npx prisma db push",
     "push:people": "cd services/people/ && npx prisma db push",
+    "push": "concurrently npm:push:*",
     "codegen:organizations": "graphql-codegen --config codegen-organizations.yml",
     "codegen:activities": "graphql-codegen --config codegen-activities.yml",
     "codegen:people": "graphql-codegen --config codegen-people.yml",
