+++ services/activities/prisma/schema.prisma
@@ -5,7 +5,7 @@ datasource db {
 
 generator client {
     provider = "prisma-client-js"
-    output   = "../../../generated/activity-db"
+    output   = "../generated/activity-db"
 }
 
 model Activity {
