+++ services/people/src/people/types.ts
@@ -1,3 +1,3 @@
-import { Person } from "../../../../generated/person-db";
+import { Person } from "../../generated/person-db";
 
 export type PersonType = Person;
